backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
PyJWT
cryptography
pydantic
httpx
